# docs/openapi/authorization-service.yaml
openapi: 3.0.3
info: {title: Aurevia Internal Authorization Service, version: 0.2.0}
security: [{workloadIdentity: []}]
paths:
...
  /internal/v1/registry/proxy-routes/{routeId}/operations:
    parameters: [{in: path, name: routeId, required: true, schema: {type: string, format: uuid}}]
    get: {operationId: listRouteOperations, parameters: [{$ref: '#/components/parameters/Actor'}], responses: {'200': {description: Route operation mappings}}}
    post: {operationId: createRouteOperation, parameters: [{$ref: '#/components/parameters/Actor'}], responses: {'201': {description: Operation created}, '400': {description: Invalid Resource/Action or pattern}}}
...
components:
  securitySchemes:
    workloadIdentity: {type: http, scheme: basic, description: Local bootstrap; production requires mTLS.}
  parameters:
    Actor: {in: header, name: X-Actor, required: true, schema: {type: string}}
...